pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli"]

[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Keyword search engine over a movie dataset using BM25 and an inverted index, with RAG-based semantic search planned as the next phase."
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "google-genai>=2.0.1",
    "nltk==3.9.1",
    "numpy==2.3.3",
    "pillow>=12.2.0",
    "progressbar>=2.5",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]

[[tool.mypy.overrides]]
module = "nltk.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "progressbar.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[tool.pylint.messages_control]
disable = ["too-few-public-methods", "duplicate-code"]

[tool.pylint.design]
max-locals = 20

[tool.pylint.typecheck]
ignored-modules = ["numpy", "sentence_transformers", "torch"]

[tool.pylint.format]
max-line-length = 88

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "furo>=2024.1.29",
    "mypy>=1.20.2",
    "myst-parser>=4.0.0",
    "pre-commit>=4.6.0",
    "pydocstyle>=6.3.0",
    "pylint>=4.0.5",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
    "sphinx>=8.0.0",
    "sphinx-autobuild>=2024.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-lint>=1.0.2",
]
